refactor(web): rename camelCase hooks and lib files to kebab-case
Brings apps/web/src files into compliance with the project
ls-lint convention. Internal-only renames; exported symbols and
the public package surface are unchanged.

Refs #162
