+++ apps/web/src/features/auth/use-auth.ts
@@ -3,8 +3,8 @@
 
 import { useContext } from 'react';
 
-import type { AuthContextValue } from './AuthContext';
-import { AuthContext } from './AuthContext';
+import type { AuthContextValue } from './auth-context';
+import { AuthContext } from './auth-context';
 
 export function useAuth(): AuthContextValue {
   const context = useContext(AuthContext);
