+++ apps/web/src/features/collection/weapons/WeaponCard.tsx
@@ -4,7 +4,7 @@
 import type { Weapon } from '@genshin/game-data';
 
 import { WeaponSummary } from '@/components/WeaponSummary';
-import { RARITY_BORDER_COLORS } from '@/lib/rarityStyles';
+import { RARITY_BORDER_COLORS } from '@/lib/rarity-styles';
 import { cn } from '@/lib/utils';
 
 interface WeaponCardProps {
