+++ apps/web/src/features/collection/weapons/use-weapon-collection-api.ts
