+++ apps/web/src/lib/rarity-styles.ts
