+++ apps/web/src/lib/weapon-types.ts
